+++ src/components/TodoSummary.tsx
@@ -1,8 +1,6 @@
-type Props = {
-  deleteAllCompleted: () => void;
-};
+import { TodoSummaryProps } from '../types/todo';
 
-export const TodoSummary = ({ deleteAllCompleted }: Props) => {
+export const TodoSummary = ({ deleteAllCompleted }: TodoSummaryProps) => {
   return (
     <div className="flex justify-end">
       <button onClick={deleteAllCompleted} className="text-sm text-red-500">
