+++ src/types/todo.ts
@@ -3,3 +3,17 @@ export type Todo = {
   title: string;
   completed: boolean;
 };
+
+export type TodoListProps = {
+  todoList: Todo[];
+  changeCompleted: (id: number) => void;
+  deleteTodo: (id: number) => void;
+};
+
+export type TodoSummaryProps = {
+  deleteAllCompleted: () => void;
+};
+
+export type AddTodoFormProps = {
+  addTodo: (title: string) => void;
+};
